move implementations of guided options into server code
